Update free-programming-playgrounds.md (#3526)
* Update free-programming-playgrounds.md

Added codesandbox.io under React (and removed the react reference from its listing on JavaScript).
Added a Java section and linked to repl.it
repl.it actually has sandboxes for most of the languages listed here, and maybe some others, but I didn't add them all in. The Swing support for Java is new, and has some bugs. They didn't respond to my report, so I think they are spread a little thin.

* add spaces to satisfy linter

* add space to satisfy linter

Co-authored-by: Eric Hellman <eric@hellman.net>